Revert "segment"
This reverts commit d2d65c32dc4261884ac8c5d07e6283513a8f8526.

+++ processing/src/main/java/org/apache/druid/query/metadata/SegmentAnalyzer.java
@@ -79,7 +79,7 @@ public class SegmentAnalyzer
     this.analysisTypes = analysisTypes;
   }
 
-  public int numRows(Segment segment)
+  public long numRows(Segment segment)
   {
     return Preconditions.checkNotNull(segment.as(PhysicalSegmentInspector.class), "PhysicalSegmentInspector")
                         .getNumRows();
